requirements.txt
apache-airflow==3.0.4
pandas==2.3.2
boto3==1.40.20